fixes https://luceeserver.atlassian.net/browse/LDEV-411

+++ core/src/main/java/lucee/runtime/ComponentImpl.java
@@ -629,7 +629,7 @@ public final class ComponentImpl extends StructSupport implements Externalizable
 		// debug yes
 		if(pc.getConfig().debug()) {
 		    DebugEntryTemplate debugEntry=pc.getDebugger().getEntry(pc,pageSource,udf.getFunctionName());//new DebugEntry(src,udf.getFunctionName());
-			int currTime=pc.getExecutionTime();
+			long currTime=pc.getExecutionTime();
 			long time=System.nanoTime();
 			
 			// sync yes
@@ -642,7 +642,7 @@ public final class ComponentImpl extends StructSupport implements Externalizable
 					}		
 					finally {
 						pc.setVariablesScope(parent);
-						int diff= ((int)(System.nanoTime()-time)-(pc.getExecutionTime()-currTime));
+						long diff= ((System.nanoTime()-time)-(pc.getExecutionTime()-currTime));
 						pc.setExecutionTime(pc.getExecutionTime()+diff);
 						debugEntry.updateExeTime(diff);	
 					}	
@@ -658,7 +658,7 @@ public final class ComponentImpl extends StructSupport implements Externalizable
 				}		
 				finally {
 					pc.setVariablesScope(parent);
-					int diff= ((int)(System.nanoTime()-time)-(pc.getExecutionTime()-currTime));
+					long diff= ((System.nanoTime()-time)-(pc.getExecutionTime()-currTime));
 					pc.setExecutionTime(pc.getExecutionTime()+diff);
 					debugEntry.updateExeTime(diff);	
 				}
